feat: finish search page
