+++ apps/website-remix/styles/app.css
@@ -20,3 +20,58 @@
 .search__poweredBy a {
   @apply absolute right-0 inline-flex items-center justify-end;
 }
+
+.ais-SearchBox-form {
+  @apply flex;
+}
+
+.ais-SearchBox-input {
+  @apply w-full py-1 pb-3 text-3xl font-light bg-transparent border-b  md:text-4xl dark:border-gray-700 focus:border-0;
+}
+
+.ais-SearchBox-submit {
+  @apply hidden;
+}
+
+.ais-SearchBox-reset {
+  @apply p-5 text-base;
+  fill: currentColor;
+}
+
+.ais-Stats {
+  @apply mt-3 mb-6 text-base font-bold;
+}
+
+/* Algolia already provides this mechanisms */
+input[type='search']::-ms-clear,
+input[type='search']::-ms-reveal {
+  @apply hidden w-0 h-0;
+}
+
+/* clears the ‘X’ from Chrome */
+input[type='search']::-webkit-search-decoration,
+input[type='search']::-webkit-search-cancel-button,
+input[type='search']::-webkit-search-results-button,
+input[type='search']::-webkit-search-results-decoration {
+  @apply hidden;
+}
+
+.ais-SearchBox-form input {
+  @apply text-xl md:text-3xl;
+  /* disable IOS native input styles */
+  /* https://stackoverflow.com/a/2918716 */
+  @apply rounded-none;
+  -webkit-appearance: none;
+}
+
+.ais-Hits-list {
+  @apply space-y-8;
+}
+
+.ais-PoweredBy {
+  @apply flex items-center justify-end font-sans text-base font-medium col-span-full;
+
+  span {
+    @apply mr-2;
+  }
+}
